+++ packages/widgetbook/lib/src/addons/inspector_addon/addon.dart
@@ -1,0 +1,1 @@
+export 'inspector_addon.dart';