+++ packages/widgetbook/lib/src/fields/field_type.dart
@@ -6,4 +6,7 @@ enum FieldType {
   list,
   string,
   duration,
+  dateTime,
+  intInput,
+  intSlider,
 }