+++ packages/widgetbook_cli/bin/main.dart
@@ -1,15 +1,17 @@
 import 'dart:io';
 
-import 'core/cli_runner.dart';
-import 'core/context_manager.dart';
+import 'package:widgetbook_cli/widgetbook_cli.dart';
 
-void main(List<String> arguments) async {
+void main(List<String> arguments, Environment environment) async {
   const contextManager = ContextManager();
-  final context = await contextManager.load(Directory.current.path);
+  final repository = await Repository.load(
+    Directory.current.path,
+  );
 
-  if (context == null) {
-    exit(1);
-  }
+  final context = await contextManager.load(
+    repository,
+    environment,
+  );
 
   await flushThenExit(
     await CliRunner(