+++ packages/widgetbook_cli/lib/src/core/environment.dart
@@ -1,0 +1,38 @@
+abstract class Environment {
+  Environment({
+    required this.name,
+    required this.appUrl,
+    required this.apiUrl,
+  });
+
+  final String name;
+  final String appUrl;
+  final String apiUrl;
+}
+
+class ProductionEnv extends Environment {
+  ProductionEnv()
+      : super(
+          name: 'production',
+          appUrl: 'https://app.widgetbook.io/#/',
+          apiUrl: 'https://api.widgetbook.io/v1/',
+        );
+}
+
+class StagingEnv extends Environment {
+  StagingEnv()
+      : super(
+          name: 'staging',
+          appUrl: 'https://dev.app.widgetbook.io/#/',
+          apiUrl: 'https://staging.api.widgetbook.io/v1/',
+        );
+}
+
+class DebugEnv extends Environment {
+  DebugEnv()
+      : super(
+          name: 'debug',
+          appUrl: 'https://dev.app.widgetbook.io/#/',
+          apiUrl: 'http://localhost:3000/v1/',
+        );
+}